+++ src/prd/mod.rs
@@ -9,6 +9,7 @@ use std::path::PathBuf;
 use structopt::StructOpt;
 
 use crate::ecoz2_lib::prd_show_file;
+use crate::lpc::lpca_cepstrum_rs::lpca_get_cepstrum;
 use crate::lpc::lpca_r_rs::lpca_r;
 
 use self::EcozPrdCommand::Show;
@@ -35,6 +36,11 @@ pub struct PrdShowOpts {
     /// Show reflection coefficients
     #[structopt(short = "k", long = "reflections")]
     show_reflections: bool,
+
+    /// Show cepstrum coefficients.
+    /// Value must be greater than the prediction order.
+    #[structopt(long = "cepstrum")]
+    show_cepstrum: Option<usize>,
 
     /// Start for coefficient range selection
     #[structopt(short = "f", long, default_value = "1")]
@@ -67,6 +73,7 @@ pub fn prd_show(opts: PrdShowOpts) -> Result<(), Box<dyn Error>> {
     let PrdShowOpts {
         show_predictors,
         show_reflections,
+        show_cepstrum,
         from,
         to,
         file,
@@ -74,7 +81,14 @@ pub fn prd_show(opts: PrdShowOpts) -> Result<(), Box<dyn Error>> {
     } = opts;
 
     if zrs {
-        prd_show_rs(file, show_predictors, show_reflections, from, to);
+        prd_show_rs(
+            file,
+            show_predictors,
+            show_reflections,
+            show_cepstrum,
+            from,
+            to,
+        );
     } else {
         prd_show_file(file, show_reflections, from, to);
     }
@@ -88,13 +102,14 @@ fn prd_show_rs(
     prd_filename: PathBuf,
     show_predictors: bool,
     show_reflections: bool,
+    show_cepstrum: Option<usize>,
     from: usize,
     to: usize,
 ) {
     let filename = prd_filename.to_str().unwrap();
     let mut prd = load(filename).unwrap();
     println!("# {}", filename);
-    prd.show(show_predictors, show_reflections, from, to);
+    prd.show(show_predictors, show_reflections, show_cepstrum, from, to);
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Debug)]
@@ -105,18 +120,38 @@ pub struct Predictor {
 }
 
 impl Predictor {
-    pub fn show(&mut self, show_predictors: bool, show_reflections: bool, from: usize, to: usize) {
+    pub fn show(
+        &mut self,
+        show_predictors: bool,
+        show_reflections: bool,
+        show_cepstrum: Option<usize>,
+        from: usize,
+        to: usize,
+    ) {
         let p = self.prediction_order;
-        let to_ = if to == 0 || to > p { p } else { to };
 
-        if show_predictors {
-            let predictors = self.get_predictors();
-            self.do_show(&predictors, "a", from, to_);
-        } else if show_reflections {
-            let reflections = self.get_reflections();
-            self.do_show(&reflections, "k", from, to_);
+        if let Some(q) = show_cepstrum {
+            if self.prediction_order < q {
+                let to_ = if to == 0 || to >= q { q - 1 } else { to };
+                let cepstrum = self.get_cepstrum(q);
+                self.do_show(&cepstrum, "c", from, to_);
+            } else {
+                eprint!(
+                    "cepstrum value must be > prediction order={}",
+                    self.prediction_order
+                );
+            }
         } else {
-            self.do_show(&self.vectors, "r", from, to_);
+            let to_ = if to == 0 || to > p { p } else { to };
+            if show_predictors {
+                let predictors = self.get_predictors();
+                self.do_show(&predictors, "a", from, to_);
+            } else if show_reflections {
+                let reflections = self.get_reflections();
+                self.do_show(&reflections, "k", from, to_);
+            } else {
+                self.do_show(&self.vectors, "r", from, to_);
+            }
         }
     }
 
@@ -153,6 +188,21 @@ impl Predictor {
             predictors.push(predictor);
         }
         predictors
+    }
+
+    fn get_cepstrum(&mut self, q: usize) -> Vec<Vec<f64>> {
+        let p = self.prediction_order;
+        assert!(p < q);
+        let mut cepstra = Vec::new();
+        let mut reflection = vec![0f64; p + 1];
+        for auto_cor in &self.vectors {
+            let mut predictor = vec![0f64; p + 1];
+            let mut cepstrum = vec![0f64; q];
+            lpca_r(p, &auto_cor, &mut reflection, &mut predictor);
+            lpca_get_cepstrum(p, auto_cor[0], &predictor, q, &mut cepstrum);
+            cepstra.push(cepstrum);
+        }
+        cepstra
     }
 
     fn get_reflections(&mut self) -> Vec<Vec<f64>> {